requirements.txt
google-cloud-aiplatform
PyMuPDF
Pillow
numpy
scikit-image
python-dotenv
requests
google-auth
tqdm